requirements.txt
Django==3.1.4
djangorestframework-simplejwt==4.4.0
djangorestframework==3.12.2
drf-yasg==1.20.0
pytest-cov==2.10.1
pytest==6.1.2
